Begin to implement the buffer

+++ src/surface.rs
@@ -1,17 +1,17 @@
 use terminal_size::{terminal_size, Height, Width};
 use std::clone::Clone;
 
-#[derive(Clone)]
+#[derive(Clone, Copy, Debug)]
 pub struct Coord {
-    pub x: i16,
-    pub y: i16,
+    pub x: usize,
+    pub y: usize,
 }
 
 impl From<(i16, i16)> for Coord {
     fn from(tuple: (i16, i16)) -> Coord {
         Coord {
-            x: tuple.0,
-            y: tuple.1,
+            x: tuple.0 as usize,
+            y: tuple.1 as usize,
         }
     }
 }
@@ -26,11 +26,12 @@ pub struct Surface {
     pub dimensions: Coord,
 }
 
+
 impl Surface {
     pub fn new() -> Surface {
         let size = terminal_size();
         if let Some((Width(w), Height(h))) = size {
-            Surface { dimensions: Coord { x: w as i16 , y: h as i16 } }
+            Surface { dimensions: Coord { x: w as usize , y: h as usize } }
         } else {
             Surface { dimensions: Coord { x: 0, y: 0} }
         }
